refactor(a11y): move XY chart axis accessibility components to chart-specific code area
