+++ packages/charts/src/components/accessibility/index.ts
@@ -8,5 +8,4 @@
 
 /* @internal */
 export { ScreenReaderSummary } from './screen_reader_summary';
-export { AxisDescriptions } from './axis_descriptions';
 export type { ChartLabelData } from './types';
